+++ android-client/app/src/main/java/de/njsm/stocks/android/dagger/modules/ActivityModule.java
@@ -31,6 +31,7 @@ import de.njsm.stocks.android.frontend.eannumber.EanNumberFragment;
 import de.njsm.stocks.android.frontend.emptyfood.EmptyFoodFragment;
 import de.njsm.stocks.android.frontend.food.FoodFragment;
 import de.njsm.stocks.android.frontend.fooditem.FoodItemFragment;
+import de.njsm.stocks.android.frontend.foodhistory.FoodHistoryFragment;
 import de.njsm.stocks.android.frontend.locations.LocationFragment;
 import de.njsm.stocks.android.frontend.main.MainActivity;
 import de.njsm.stocks.android.frontend.main.OutlineFragment;
@@ -90,6 +91,9 @@ public abstract class ActivityModule {
 
     @ContributesAndroidInjector
     abstract ShoppingListFragment contributeShoppingListFragment();
+
+    @ContributesAndroidInjector
+    abstract FoodHistoryFragment contributeFoodHistoryFragment();
 
     @ContributesAndroidInjector
     abstract MainActivity contributeMainActivity();